[WIP] Working at new login screen. Errors interface extracted to individual files
