+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/errors/IsValidEmail.kt
@@ -1,0 +1,7 @@
+package com.softyorch.taskapp.ui.screensBeta.login.errors
+
+import android.util.Patterns
+
+interface IsValidEmail {
+    fun isValidEmail(email: String): Boolean = Patterns.EMAIL_ADDRESS.matcher(email).matches()
+}
