+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/errors/IsValidEmailRepeat.kt
@@ -1,0 +1,5 @@
+package com.softyorch.taskapp.ui.screensBeta.login.errors
+
+interface IsValidEmailRepeat {
+    fun isValidEmail(email: String, emailRepeat: String): Boolean = (email == emailRepeat)
+}
