+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/errors/IsValidName.kt
@@ -1,0 +1,5 @@
+package com.softyorch.taskapp.ui.screensBeta.login.errors
+
+interface IsValidName {
+    fun isValidName(name: String): Boolean = (name.length >= 3)
+}
