+++ app/src/main/java/com/softyorch/taskapp/ui/screensBeta/login/errors/IsValidPassRepeat.kt
@@ -1,0 +1,5 @@
+package com.softyorch.taskapp.ui.screensBeta.login.errors
+
+interface IsValidPassRepeat {
+    fun isValidPass(pass: String, passRepeat: String): Boolean = (pass == passRepeat)
+}
